requirements.txt
cbor2==5.2.0
certifi==2020.12.5
chardet==4.0.0
Deprecated==1.2.12
idna==2.10
luma.core==2.3.1
luma.emulator==1.4.0
Pillow==8.1.2
pyftdi==0.52.9
pygame==2.0.1
pyserial==3.5
pytz==2021.1
pyusb==1.1.1
requests==2.25.1
smbus2==0.4.1
urllib3==1.26.4
wrapt==1.12.1
